Remove non-required packages; refactor helpers
