+++ lib/jsdom/living/xhr/xhr-utils.js
@@ -348,7 +348,6 @@ function createClient(xhr) {
   } else if (hasBody && getRequestHeader(flag.requestHeaders, "content-type") === null) {
     requestHeaders["Content-Type"] = "text/plain;charset=UTF-8";
   }
-
   const options = {
     agents,
     method: flag.method,
@@ -368,7 +367,7 @@ function createClient(xhr) {
   function doRequest(sendAuthImmediately = false) {
     try {
       // set authentication header if sendAuthImmediately is set
-      const clientRequest = request(uri, options);
+      const clientRequest = request(urlObj, options);
       if (hasBody && flag.formData) {
         form.removeAllListeners();
         form.on("error", err => {
@@ -425,7 +424,7 @@ function createClient(xhr) {
       };
       client.headers = clientRequest._options.headers;
       client.emit("request", clientRequest._currentRequest);
-      clientRequest.end();
+      process.nextTick(() => clientRequest.end());
     } catch (e) {
       process.nextTick(() => client.emit("error", e));
     }
@@ -433,7 +432,6 @@ function createClient(xhr) {
 
   const nonSimpleHeaders = Object.keys(flag.requestHeaders)
     .filter(header => !simpleHeaders.has(header.toLowerCase()));
-
   if (crossOrigin && (!simpleMethods.has(ucMethod) || nonSimpleHeaders.length > 0 || properties.uploadListener)) {
     const preflightRequestHeaders = [];
     for (const header in requestHeaders) {
@@ -452,7 +450,6 @@ function createClient(xhr) {
     preflightRequestHeaders["User-Agent"] = flag.userAgent;
 
     flag.preflight = true;
-
     const preflightOptions = {
       agents,
       method: "OPTIONS",
@@ -460,7 +457,7 @@ function createClient(xhr) {
       followRedirects: false
     };
 
-    const preflightClient = request(uri, preflightOptions);
+    const preflightClient = request(urlObj, preflightOptions);
 
     preflightClient.on("response", resp => {
       // don't send the real request if the preflight request returned an error
@@ -481,7 +478,7 @@ function createClient(xhr) {
     client.abort = () => {
       preflightClient.abort();
     };
-    preflightClient.end();
+    process.nextTick(() => preflightClient.end());
   } else {
     doRequest();
   }
