+++ package.json
@@ -32,16 +32,13 @@
     "escodegen": "^2.0.0",
     "fetch-cookie": "^0.11.0",
     "form-data": "^3.0.0",
-    "got": "^11.8.2",
     "html-encoding-sniffer": "^2.0.1",
     "http-proxy-agent": "^4.0.1",
     "https-proxy-agent": "^5.0.0",
     "is-potential-custom-element-name": "^1.0.0",
-    "node-fetch": "^2.6.0",
+    "node-fetch": "^2.6.1",
     "nwsapi": "^2.2.0",
     "parse5": "6.0.1",
-    "request": "^2.88.2",
-    "request-promise-native": "^1.0.9",
     "saxes": "^5.0.1",
     "symbol-tree": "^3.2.4",
     "tough-cookie": "^4.0.0",
